PictureView
In the toolbox pick picture-box , Table layer out, Flow layer out name according to the name it suit you, pick 6 buttons name one to load picture, name the second for background color, name the third for rotating the picture, fourth for picture background ,fifth for zoon in, six for zoom out. After finishing , click each button , copy the code and paste, repeat it for the rest of  buttons, hope you enjoy it.